Fixed Bug#2
Fixed Bug#2 - The second bug is the problem with pieces appearing in the drop zones on reset / choosing a new puzzle. Those should be removed / reparented back to the drag zone as well, so that the player has a fresh board to drop onto.

I think I wasn't fully able to fix the 2nd bug because when ever it resets the puzzles disappear